+++ src/commands/git/deleteProductionReleaseBranch.ts
@@ -18,6 +18,7 @@ const deleteProductionReleaseBranch = (context: vscode.ExtensionContext, outputC
     }
     const gitRepo = vscode.workspace.workspaceFolders[0].uri.fsPath;
     const git = new Git(gitRepo, (context.workspaceState.get("t00ls.mode") as t00lsMode));
+    await git.initialize();
 
     // fetch the latest updates from remote
     try {
@@ -63,18 +64,19 @@ const deleteProductionReleaseBranch = (context: vscode.ExtensionContext, outputC
     if (!branch) { throw new Error("Error selecting production release branch..."); }
 
     // Before deleting a production release branch, check to see if
-    // the branch has been merged into master.
-    const numCommitsAheadOfMaster = await git.getNumberOfCommitsAheadOfBranch("master", selected.label);
-    if (numCommitsAheadOfMaster > 0) {
-      if (!(await promptYesNo({ question: `There are commits in '${selected.label}' that do not exist in 'master'. Delete?`, noIsDefault: true, ignoreFocusOut: true }))) { return; }
+    // the branch has been merged into the main branch.
+    const mainBranchName = await git.getMainBranchName();
+    const numCommitsAheadOfMain = await git.getNumberOfCommitsAheadOfBranch(mainBranchName, selected.label);
+    if (numCommitsAheadOfMain > 0) {
+      if (!(await promptYesNo({ question: `There are commits in '${selected.label}' that do not exist in '${mainBranchName}'. Delete?`, noIsDefault: true, ignoreFocusOut: true }))) { return; }
     }
 
     if (branch.name === currentBranch) {
-      // switch to master
+      // switch to the main branch
       try {
-        await git.checkoutBranch("master");
+        await git.checkoutBranch(mainBranchName);
       } catch (e) {
-        showErrorMessage(outputChannel, `There was an error switching to 'master': ${e}`);
+        showErrorMessage(outputChannel, `There was an error switching to '${mainBranchName}': ${e}`);
         return;
       }
     }