+++ src/commands/git/syncRepo.ts
@@ -20,6 +20,7 @@ const syncRepo = (context: vscode.ExtensionContext, outputChannel: vscode.Output
     const gitRepo = vscode.workspace.workspaceFolders[0].uri.fsPath;
     const mode: t00lsMode = context.workspaceState.get("t00ls.mode") as t00lsMode;
     const git = new Git(gitRepo, mode);
+    await git.initialize();
     const relationshipCache = BranchRelationshipCache.getInstance();
 
     if (mode === t00lsMode.Normal) {
@@ -78,30 +79,31 @@ const syncRepo = (context: vscode.ExtensionContext, outputChannel: vscode.Output
       return;
     }
 
-    // switch to master
-    if (currentBranch !== "master") {
+    // switch to the main branch.
+    const mainBranchName = await git.getMainBranchName();
+    if (currentBranch !== mainBranchName) {
       try {
-        await git.checkoutBranch("master");
+        await git.checkoutBranch(mainBranchName);
       } catch (e) {
-        showErrorMessage(outputChannel, `Error switching to 'master': ${e}`);
+        showErrorMessage(outputChannel, `Error switching to '${mainBranchName}': ${e}`);
         return;
       }
     }
 
     if (mode === t00lsMode.Normal) {
-      // pull master
+      // pull the main branch
       try {
         await git.pull();
       } catch (e) {
-        showErrorMessage(outputChannel, `Error pulling to 'master': ${e}`);
+        showErrorMessage(outputChannel, `Error pulling to '${mainBranchName}': ${e}`);
         return;
       }
 
-      // push master
+      // push the main branch
       try {
         await git.push();
       } catch (e) {
-        showErrorMessage(outputChannel, `Error pushing to remote 'master': ${e}`);
+        showErrorMessage(outputChannel, `Error pushing to remote '${mainBranchName}': ${e}`);
         return;
       }
     }
@@ -128,11 +130,11 @@ const syncRepo = (context: vscode.ExtensionContext, outputChannel: vscode.Output
         }
       }
 
-      // merge master into the production release branch
+      // merge the main branch into the production release branch
       try {
-        await git.mergeBranch("master");
+        await git.mergeBranch(mainBranchName);
       } catch (e) {
-        showErrorMessage(outputChannel, `Error merging 'master' into '${prodRelease.name}': ${e}`);
+        showErrorMessage(outputChannel, `Error merging '${mainBranchName}' into '${prodRelease.name}': ${e}`);
         return;
       }
 
@@ -146,7 +148,7 @@ const syncRepo = (context: vscode.ExtensionContext, outputChannel: vscode.Output
             const remotes = await git.getAllRemotes();
             let remote: string | undefined = remotes[0];
             if (remotes.length > 1) {
-              remote = await vscode.window.showQuickPick(remotes, {});
+              remote = await vscode.window.showQuickPick(remotes, { ignoreFocusOut: true });
               if (!remote) { return; }
             }
 
@@ -192,21 +194,44 @@ const syncRepo = (context: vscode.ExtensionContext, outputChannel: vscode.Output
         }
       }
 
-      let baseBranch: vscode.QuickPickItem | undefined = { label: relationshipCache.getRelationship(featureBranch.name)?.productionReleaseBranch || "master", description: "Local" };
-      let baseBranches = [{ label: "master", description: "Local" }];
-
-      if (productionReleaseBranches.length > 0) {
-        baseBranches = baseBranches.concat(productionReleaseBranches.map(p => ({ label: p.name, description: (p.remote) ? "Remote" : "Local" })));
-        baseBranch = await vscode.window.showQuickPick(baseBranches, { canPickMany: false, placeHolder: `What is the base branch for feature branch '${featureBranch.name}'?`, ignoreFocusOut: true });
-        if (!baseBranch) {
-          baseBranch = { label: "master", description: "Local" };
+      let savedRelationship = relationshipCache.getRelationship(featureBranch.name);
+      if (savedRelationship) {
+        // verify the production release branch still exists locally
+        if (_.find(productionReleaseBranches, { name: savedRelationship.productionReleaseBranch }) === undefined) {
+          // it doesn't exist, so delete all relationships related to the production release branch
+          // clear the relationship for the feature branch as well.
+          await relationshipCache.clearRelationshipsForProductionReleaseBranch(savedRelationship.productionReleaseBranch);
+          await relationshipCache.clearRelationshipForFeatureBranch(featureBranch.name);
+          savedRelationship = undefined;
         }
       }
 
-      // ask if the user wants to save the production release branch/feature branch relationship for next time.
-      if (baseBranch.label !== "master" && (await promptYesNo({ question: "Save this relationship for future syncs?" }))) {
-        await relationshipCache.setRelationship(featureBranch.name, baseBranch.label);
+      let baseBranches: vscode.QuickPickItem[] = [{ label: mainBranchName, description: "Local" }];
+      let baseBranch: vscode.QuickPickItem | undefined = baseBranches[0];
+      if (productionReleaseBranches.length > 0) {
+        baseBranches = baseBranches.concat(productionReleaseBranches.map(p => ({ label: p.name, description: (p.remote) ? "Remote" : "Local" })));
       }
+
+      // if the base branch isn't saved, and there is more than one base branch to select from,
+      // show a prompt to the user
+      if (baseBranch.label === mainBranchName && baseBranches.length > 1 && savedRelationship === undefined) {
+        baseBranch = await vscode.window.showQuickPick(baseBranches, { canPickMany: false, placeHolder: `What is the base branch for feature branch '${featureBranch.name}'?`, ignoreFocusOut: true });
+        if (!baseBranch) {
+          return;
+        }
+      }
+
+      if (savedRelationship === undefined) {
+        // ask if the user wants to save the production release branch/feature branch relationship for next time.
+        if (baseBranch.label !== mainBranchName && (await promptYesNo({ question: "Save this relationship for future syncs?" }))) {
+          await relationshipCache.setRelationship(featureBranch.name, baseBranch.label);
+        }
+      } else {
+        // use the production release branch defined by the saved relationship 
+        // as the base branch for the production release branch.
+        baseBranch.label = savedRelationship.productionReleaseBranch;
+      }
+      
 
       if ((await promptYesNo({ question: `Merge '${baseBranch.label}' into '${featureBranch.name}'` }))) {
         // merge the base branch into the feature branch
@@ -263,15 +288,15 @@ const syncRepo = (context: vscode.ExtensionContext, outputChannel: vscode.Output
         return;
       }
     } else {
-      // the branch must've been deleted. Switch to master.
+      // the branch must've been deleted. Switch to the main branch.
       try {
-        await git.checkoutBranch("master");
+        await git.checkoutBranch(mainBranchName);
         // ask about poping stash
         if (await promptYesNo({ question: `'${currentBranch}' was deleted. Do you want to pop the stash that was created anyways?` })) {
           await git.popStash();
         }
       } catch (e) {
-        showErrorMessage(outputChannel, `Error switching to 'master': ${e}`);
+        showErrorMessage(outputChannel, `Error switching to '${mainBranchName}': ${e}`);
         return;
       }
     }