+++ src/types/NeedsAsyncInitialization.ts
@@ -1,0 +1,6 @@
+interface NeedsAsyncInitialization {
+
+  initialize(...params: any): Promise<void>;
+}
+
+export default NeedsAsyncInitialization;