+++ src/types/NeedsSyncInitialization.ts
@@ -1,0 +1,6 @@
+interface NeedsSyncInitialization {
+
+  initialize(...params: any): void;
+}
+
+export default NeedsSyncInitialization;